diskmon/requirements.txt
pyYAML
pySMART.smartx==0.3.9
twilio==6.22.1
